src/test/java/org/alan/algorithm/practice/baekjoon/silver/three/BJ1935Test.java
package org.alan.algorithm.practice.baekjoon.silver.three;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ1935Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        5
                        ABC*+DE/-
                        1
                        2
                        3
                        4
                        5""", "6.20"),
                Arguments.of("""
                        1
                        AA+A+
                        1""", "3.00")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ1935().solve(input));
    }
}
